requirements.txt
requests==2.31.0
openai==1.30.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
httpx<0.28.0
